+++ SapSecurity/SapSecurity/Services/Db/ISensorLogService.cs
@@ -4,7 +4,7 @@ namespace SapSecurity.Services.Db;
 
 public interface ISensorLogService
 {
-    Task LogAsync(double status, int sensorDetailId);
+    Task LogAsync(int status, int sensorDetailId);
 
     Task SaveChangesAsync();
     Task<SensorLog?> GetLastLogAsync(int sensorId);
